build_dataset.py: added possibility to have a reference dataset that has been split along templates, adapted the class overall and more specifically the split_dataset method. balancing_workflow.py: added the split_by_template_balanced method, which allows to split a dataframe into train/test/valid sets along templates while respecting the relative proportions. run_build_dataset.py: has been adapted for the new argument alllowing to say that a dataframe is split by templates already.

+++ reaction_equilibration/uspto_balance/src/uspto_balance/build_dataset.py
@@ -13,16 +13,18 @@ class BuildDataset:
                  target_rxns_per_template, 
                  path_to_save_subset_split, 
                  train_size, 
-                 skip_appending_rxns_part=False):
+                 skip_appending_rxns_part=False, 
+                 is_ref_dataset_split_by_temp=False):
         
-        self.path_to_folder             = path_to_folder
-        self.dataset_name               = dataset_name
-        self.df_temp_path               = df_temp_path
-        self.path_to_save_full_csv      = path_to_save_full_csv
-        self.path_to_ref_dataset        = path_to_ref_dataset
-        self.target_rxns_per_template   = target_rxns_per_template
-        self.train_size                 = train_size
-        self.path_to_save_subset_split  = path_to_save_subset_split
+        self.path_to_folder               = path_to_folder
+        self.dataset_name                 = dataset_name
+        self.df_temp_path                 = df_temp_path
+        self.path_to_save_full_csv        = path_to_save_full_csv
+        self.path_to_ref_dataset          = path_to_ref_dataset
+        self.target_rxns_per_template     = target_rxns_per_template
+        self.train_size                   = train_size
+        self.path_to_save_subset_split    = path_to_save_subset_split
+        self.is_ref_dataset_split_by_temp = is_ref_dataset_split_by_temp
 
         # initialize DataHandler
         datahandler = DataHandler(file_path=self.df_temp_path)
@@ -33,7 +35,9 @@ class BuildDataset:
             self.df = datahandler.load_data()
 
     def append_rxns_to_csv(self):
-        _ = bw.append_saved_rxns_into_csv_old(self.path_to_folder, self.dataset_name, self.df_temp, self.path_to_save_full_csv)
+        #_ = bw.append_saved_rxns_into_csv_old(self.path_to_folder, self.dataset_name, self.df_temp, self.path_to_save_full_csv)
+        _ = bw.append_saved_rxns_into_csv(self.path_to_folder, self.dataset_name, self.df_temp, self.path_to_save_full_csv)
+
 
     def load_conf_rxns_df(self):
         # initialize datahandler class
@@ -45,28 +49,28 @@ class BuildDataset:
 
     def drop_rxns_in_common_from_df(self):
         # create a column 'isin_original_set' in the dataframe with boolean values
-        print('self.df line 48:', self.df)
-        print('self.df_ref_data line 49:', self.df_ref_data)
         self.df = bw.label_reactions_from_original_dataframe(self.df, original_rxns=self.df_ref_data['Reaction'], reaction_col='rxns')
-        print('self.df line 51:', self.df)
         # drop the reactions that are in the original dataset
         self.df = self.df[self.df['isin_original_set']==False].reset_index(drop=True)
-        print('self.df line 54:', self.df)
         # drop the column 'isin_original_set'
-        #df = self.df.drop(columns=['isin_original_set'], inplace=True)
         return self.df.drop(columns=['isin_original_set'], inplace=False)
-        #print('print df line 57:', df)        
-        #return df
+
     
     def create_dict_of_trainorvalid_templates(self, df, ref_data):
         # get a dictionary of all the templates that are in 'TRAIN' and 'VALID' sets
-        print('Getting all the templates in the reference dataset that are in the TRAIN and VALID sets...')
-        print('ref_data line 63: ', ref_data)
-        print('df line 60:', df)
-        trainorvalid_uspto_df = ref_data[ref_data['Set'].isin(['TRAIN', 'VALID'])]
+        print('Getting all the templates in the reference dataset that are in the TRAIN and VAL sets...')
+        trainorvalid_uspto_df = ref_data[ref_data['Set'].isin(['TRAIN', 'VAL'])]
         trainorvalid_templates = df[df['retro_template'].isin(trainorvalid_uspto_df['retro_template_r1_full_corr'])]
         trainorvalid_template_dict = {template: True for template in trainorvalid_templates['retro_template'].unique()}
         return trainorvalid_template_dict
+    
+    def create_dict_of_split_set_templates(self, df, ref_data, split_set:str):
+        # get a dictionary of all the templates that are in the 'split_set' set
+        print(f'Getting all the templates in the reference dataset that are in the {split_set} set...')
+        split_set_ref_data = ref_data[ref_data['Set'].isin([f'{split_set}'])]
+        split_set_templates = df[df['retro_template'].isin(split_set_ref_data['retro_template_r1_full_corr'])]
+        split_set_templates_dict = {template: True for template in split_set_templates['retro_template'].unique()}
+        return split_set_templates_dict
     
     def assign_test_trainvalid_sets(self, trainorvalid_template_dict):
             # assign TEST split to all the reactions
@@ -83,6 +87,20 @@ class BuildDataset:
 
             return self.testset_df, self.trainvalidset_df
 
+    def assign_test_train_valid_sets_easycase(self, train_dict, test_dict, valid_dict):
+        # make sure the sets are non assigned for last step
+        self.df['Set'] = ''
+        self.df.loc[self.df['retro_template'].isin(train_dict.keys()), 'Set'] = 'TRAIN'
+        self.df.loc[self.df['retro_template'].isin(test_dict.keys()), 'Set'] = 'TEST'
+        self.df.loc[self.df['retro_template'].isin(valid_dict.keys()), 'Set'] = 'VALID'
+        # assign the non-assigned templates to the training set
+        self.df.loc[self.df['Set']=='', 'Set'] = 'TRAIN'
+
+        self.trainset_df = self.df[self.df['Set']=='TRAIN']
+        self.testset_df = self.df[self.df['Set']=='TEST']
+        self.validset_df = self.df[self.df['Set']=='VALID']
+
+        return self.trainset_df, self.testset_df, self.validset_df
 
     def get_dataset(self):
         # save all csv files into a single one containing all the valid reactions
@@ -102,7 +120,7 @@ class BuildDataset:
     
     def split_dataset(self):
         # if a reference dataset exists, load it and use it to assign the splits to the newly create dataset?
-        if os.path.exists(self.path_to_ref_dataset):
+        if os.path.exists(self.path_to_ref_dataset) and not self.is_ref_dataset_split_by_temp:
             # load the reference dataset
             datahandler_refdataset = DataHandler(file_path=self.path_to_ref_dataset)
             print('Loading reference dataset...')
@@ -111,9 +129,8 @@ class BuildDataset:
             # remove the reactions in the created dataset already in the reference dataset
             # -- create a column 'isin_original_set' in the dataframe with boolean values
             print('Removing reactions that are also present in the original dataset...')
-            print('self.df line 107: ',self.df)
             self.df = self.drop_rxns_in_common_from_df()
-            print('self.df line 109: ',self.df)
+            
             # get a dictionary of all the templates that are in 'TRAIN' and 'VALID' sets
             trainorvalid_template_dict = self.create_dict_of_trainorvalid_templates(df=self.df, ref_data=self.df_ref_data)
 
@@ -127,15 +144,59 @@ class BuildDataset:
             print('Splitting the train/validation set into TRAIN and VALID sets...')
             self.trainvalidset_df = self.trainvalidset_df.groupby('retro_template', group_keys=False).apply(lambda x: bw.stratified_split(x, self.train_size))
 
+            # keep maximum 20 reactions per template from the test set
+            testset_20RpT = bw.subset_n_random_reactions_per_template(self.testset_df, 20)
             # concat the full test set and the train/valid subsets
             print('Concatenating the test set and the train/validation set...')
-            self.df = pd.concat([self.trainvalidset_df, self.testset_df])
+            self.df = pd.concat([self.trainvalidset_df, testset_20RpT])
+
+            print('Saving the final dataset...')
+            self.df.to_csv(self.path_to_save_subset_split)
+
+        elif os.path.exists(self.path_to_ref_dataset) and self.is_ref_dataset_split_by_temp:
+            # load the reference dataset
+            datahandler_refdataset = DataHandler(file_path=self.path_to_ref_dataset)
+            print('Loading reference dataset...')
+            self.df_ref_data = datahandler_refdataset.load_data()
+
+            # remove the reactions in the created dataset already in the reference dataset
+            # -- create a column 'isin_original_set' in the dataframe with boolean values
+            print('Removing reactions that are also present in the original dataset...')
+            self.df = self.drop_rxns_in_common_from_df()    
+
+            # get a dictionary of the templates in 'TRAIN' set
+            train_template_dict = self.create_dict_of_split_set_templates(df=self.df, ref_data=self.df_ref_data, split_set='TRAIN')
+            # get a dictionary of the templates in 'VALID' set
+            valid_template_dict = self.create_dict_of_split_set_templates(df=self.df, ref_data=self.df_ref_data, split_set='VALID')
+            # get a dictionary of the templates in 'TEST' set
+            test_template_dict  = self.create_dict_of_split_set_templates(df=self.df, ref_data=self.df_ref_data, split_set='TEST')
+
+            # Define 'TEST', 'VALID' and 'TRAIN' sets
+            self.trainset_df, self.testset_df, self.validset_df = self.assign_test_train_valid_sets_easycase(train_dict=train_template_dict, 
+                                                                                                             test_dict=test_template_dict, 
+                                                                                                             valid_dict=valid_template_dict)
+            
+            # take self.target_rxns_per_template reactions per template in the split sets
+            print(f'Subsetting {self.target_rxns_per_template} reactions per template in the training set...')
+            self.trainset_df = bw.subset_n_random_reactions_per_template(self.trainset_df, target=self.target_rxns_per_template)            
+            print(f'Subsetting {self.target_rxns_per_template} reactions per template in the test set...')
+            self.testset_df = bw.subset_n_random_reactions_per_template(self.testset_df, target=self.target_rxns_per_template) 
+            print(f'Subsetting {self.target_rxns_per_template} reactions per template in the validation set...')
+            self.validset_df = bw.subset_n_random_reactions_per_template(self.validset_df, target=self.target_rxns_per_template)
+
+            # concat the full test set and the train/valid subsets
+            print('Concatenating the training, test, and validation sets...')
+            self.df = pd.concat([self.trainset_df, self.testset_df, self.validset_df])
 
             print('Saving the final dataset...')
             self.df.to_csv(self.path_to_save_subset_split)
 
         # if there is no reference dataset, split the dataset into 'TRAIN', 'TEST', and 'VAL' sets according to templates
         else:
-            self.df = self.df.groupby('retro_template', group_keys=False).apply(lambda x: bw.stratified_train_val_test_split(x, self.train_size))
+            # in case a stratified split is wanted
+            #self.df = self.df.groupby('retro_template', group_keys=False).apply(lambda x: bw.stratified_train_val_test_split(x, self.train_size))
+            #self.df.to_csv(self.path_to_save_subset_split)
+            
+            # in case a complete separation of templates between the splits is wanted
+            self.df = bw.split_by_template_balanced(df=self.df, template_col='retro_template_r1_full_corr')
             self.df.to_csv(self.path_to_save_subset_split)
-
